Still refactoring the commands
