+++ usaspending_api/etl/es_etl_helpers.py
@@ -1,12 +1,17 @@
-import subprocess
-from datetime import datetime
+import boto3
 import csv
 import os
-from time import perf_counter, sleep
+import json
 import pandas as pd
+import pytz
+import subprocess
+import tempfile
 
-from elasticsearch import helpers
+from collections import defaultdict
+from datetime import datetime
 from django.db import connection
+from elasticsearch import helpers
+from time import perf_counter, sleep
 
 AWARD_DESC_CATEGORIES = {
     'loans': 'loans',
@@ -16,6 +21,8 @@ AWARD_DESC_CATEGORIES = {
     'contract': 'contracts',
     'direct payment': 'directpayments'
 }
+
+UNIVERSAL_TRANSACTION_ID_NAME = 'transaction_id'
 
 
 class DataJob:
@@ -140,6 +147,25 @@ def csv_chunk_gen(filename, header, chunksize, job_id):
         yield file_df.to_dict(orient='records')
 
 
+def es_data_loader(client, fetch_jobs, done_jobs, config):
+    while True:
+        if not done_jobs.empty():
+            job = done_jobs.get_nowait()
+            if job.name is None:
+                break
+
+            printf({'msg': 'Starting new job', 'job': job.name, 'f': 'ES Ingest'})
+            post_to_elasticsearch(client, job, config)
+            if os.path.exists(job.csv) and not config['keep']:
+                os.remove(job.csv)
+        else:
+            printf({'msg': 'No Job :-( Sleeping 15s', 'f': 'ES Ingest'})
+            sleep(15)
+
+    printf({'msg': 'Completed Elasticsearch data load', 'f': 'ES Ingest'})
+    return
+
+
 def streaming_post_to_es(client, chunk, index_name, job_id=None):
     success, failed = 0, 0
     try:
@@ -172,7 +198,19 @@ def post_to_elasticsearch(client, job, config, chunksize=250000):
 
     csv_generator = csv_chunk_gen(job.csv, VIEW_COLUMNS, chunksize, job.name)
     for count, chunk in enumerate(csv_generator):
+        if len(chunk) == 0:
+            printf({'msg': 'No records to add or delete for chunk {}'.format(count), 'f': 'ES Ingest', 'job': job.name})
+            continue
         iteration = perf_counter()
+        if config['recreate'] is False:
+            id_list = [{'key': c[UNIVERSAL_TRANSACTION_ID_NAME], 'col': UNIVERSAL_TRANSACTION_ID_NAME} for c in chunk]
+            delete_transactions_from_es(client, id_list, job.name, config, job.index)
+
+        printf({
+            'msg': 'streaming chunk {} to ES'.format(count),
+            'job': job.name,
+            'f': 'ES Ingest'
+        })
         streaming_post_to_es(client, chunk, job.index, job.name)
         printf({
             'msg': 'Iteration on chunk {} took {}s'.format(count, perf_counter() - iteration),
@@ -207,9 +245,160 @@ def csv_row_count(filename, has_header=True):
     return row_count
 
 
+def gather_deleted_ids(config):
+    '''
+    Connect to S3 and gather all of the transaction ids stored in CSV files
+    generated by the broker when transactions are removed from the DB.
+    '''
+
+    if not config['provide_deleted']:
+        printf({'msg': 'Skipping the S3 CSV fetch for deleted transactions'})
+        return
+    printf({'msg': 'Gathering all deleted transactions from S3'})
+    start = perf_counter()
+    try:
+        s3 = boto3.resource('s3', region_name=config['aws_region'])
+        bucket = s3.Bucket(config['s3_bucket'])
+        bucket_objects = list(bucket.objects.all())
+    except Exception as e:
+        print('\n[ERROR]\n')
+        print('Verify settings.CSV_AWS_REGION and settings.DELETED_TRANSACTIONS_S3_BUCKET_NAME are correct')
+        print('  or is using env variables: CSV_AWS_REGION and DELETED_TRANSACTIONS_S3_BUCKET_NAME')
+        print('\n{}\n'.format(e))
+        raise SystemExit
+
+    if config['verbose']:
+        printf({'msg': 'CSV data from {} to now'.format(config['starting_date'])})
+
+    to_datetime = datetime.combine(config['starting_date'], datetime.min.time(), tzinfo=pytz.UTC)
+    filtered_csv_list = [
+        x for x in bucket_objects
+        if (
+            x.key.endswith('.csv') and
+            not x.key.startswith('staging') and
+            x.last_modified >= to_datetime
+        )
+    ]
+
+    if config['verbose']:
+        printf({'msg': 'Found {} csv files'.format(len(filtered_csv_list))})
+
+    deleted_ids = {}
+
+    for obj in filtered_csv_list:
+        # Use temporary files to facilitate date moving from csv files on S3 into pands
+        (file, file_path) = tempfile.mkstemp()
+        bucket.download_file(obj.key, file_path)
+
+        # Ingests the CSV into a dataframe. pandas thinks some ids are dates, so disable parsing
+        data = pd.DataFrame.from_csv(file_path, parse_dates=False)
+
+        if 'detached_award_proc_unique' in data:
+            new_ids = ['cont_tx_' + x for x in data['detached_award_proc_unique'].values]
+        elif 'afa_generated_unique' in data:
+            new_ids = ['asst_tx_' + x for x in data['afa_generated_unique'].values]
+        else:
+            printf({'msg': '  [Missing valid col] in {}'.format(obj.key)})
+
+        # Next statements are ugly, but properly handle the temp files
+        os.close(file)
+        os.remove(file_path)
+
+        if config['verbose']:
+            printf({'msg': '{:<55} ... {}'.format(obj.key, len(new_ids))})
+
+        for uid in new_ids:
+            if uid in deleted_ids:
+                if deleted_ids[uid]['timestamp'] < obj.last_modified:
+                    deleted_ids[uid]['timestamp'] = obj.last_modified
+            else:
+                deleted_ids[uid] = {'timestamp': obj.last_modified}
+
+    if config['verbose']:
+        for k, v in deleted_ids.items():
+            printf({'msg': 'id: {} last modified: {}'.format(k, str(v['timestamp']))})
+
+    printf({'msg': 'Gathering {} deleted transactions took {}s'.format(len(deleted_ids), perf_counter() - start)})
+    return deleted_ids
+
+
+def filter_query(column, values, query_type="match_phrase"):
+    queries = [{query_type: {column: str(i)}} for i in values]
+    body = {
+        "query": {
+            "bool": {
+                "should": [
+                    queries
+                ]
+            }
+        }
+    }
+
+    return json.dumps(body)
+
+
+def delete_query(response):
+    body = {
+        "query": {
+            "ids": {
+                "type": "transaction_mapping",
+                "values": [i['_id'] for i in response['hits']['hits']]
+            }
+        }
+    }
+    return json.dumps(body)
+
+
+def chunks(l, n):
+    """Yield successive n-sized chunks from l."""
+    for i in range(0, len(l), n):
+        yield l[i:i + n]
+
+
+def delete_transactions_from_es(client, id_list, job_id, config, index=None, size=50000):
+    '''
+    id_list = [{key:'key1',col:'tranaction_id'},
+               {key:'key2',col:'generated_unique_transaction_id'}],
+               ...]
+    '''
+    start = perf_counter()
+
+    printf({
+        'msg': 'Deleting up to {} transaction(s)'.format(len(id_list)),
+        'f': 'ES Delete',
+        'job': job_id})
+
+    if index is None:
+        index = '{}-*'.format(config['root_index'])
+    start_ = client.search(index=index)['hits']['total']
+    printf({
+        'msg': 'Starting amount of indices ----- {}'.format(start_),
+        'f': 'ES Delete',
+        'job': job_id})
+    col_to_items_dict = defaultdict(list)
+    for l in id_list:
+        col_to_items_dict[l['col']].append(l['key'])
+
+    for column, values in col_to_items_dict.items():
+        printf({'msg': 'Deleting from col {}'.format(column), 'f': 'ES Delete', 'job': job_id})
+        values_generator = chunks(values, 1000)
+        for values_ in values_generator:
+            body = filter_query(column, values_)
+            response = client.search(index=index, body=body, size=size)
+            delete_body = delete_query(response)
+            client.delete_by_query(index=index, body=delete_body, size=size)
+    end_ = client.search(index=index)['hits']['total']
+
+    t = perf_counter() - start
+    total = str(start_ - end_)
+    printf({'msg': 'ES Deletes took {}s. Deleted {} records'.format(t, total), 'f': 'ES Delete', 'job': job_id})
+    return
+
 # ==============================================================================
 # SQL Template Strings for Postgres Statements
 # ==============================================================================
+
+
 VIEW_COLUMNS = [
     'transaction_id', 'modification_number', 'award_id', 'piid', 'fain', 'uri',
     'award_description', 'product_or_service_code', 'product_or_service_description',
@@ -257,6 +446,8 @@ SELECT transaction_id, generated_unique_transaction_id, update_date FROM transac
 WHERE EXISTS (
   SELECT *
   FROM temp_transaction_ids
-  WHERE transaction_delta_view.generated_unique_transaction_id = temp_transaction_ids.generated_unique_transaction_id and transaction_fiscal_year={fy}
+  WHERE
+    transaction_delta_view.generated_unique_transaction_id = temp_transaction_ids.generated_unique_transaction_id
+    AND transaction_fiscal_year={fy}
 );
 '''
