+++ usaspending_api/etl/management/commands/fetch_transactions.py
@@ -1,9 +1,5 @@
-import boto3
 import os
-import pandas as pd
-import pytz
 import subprocess
-import tempfile
 
 from datetime import date
 from datetime import datetime
@@ -12,6 +8,7 @@ from time import perf_counter
 
 from usaspending_api.etl.es_etl_helpers import configure_sql_strings
 from usaspending_api.etl.es_etl_helpers import execute_sql_statement
+from usaspending_api.etl.es_etl_helpers import gather_deleted_ids
 
 # SCRIPT OBJECTIVES and ORDER OF EXECUTION STEPS
 # 1. [conditional] Gather the list of deleted transactions from S3
@@ -185,81 +182,3 @@ def set_config():
     }
 
     return config
-
-
-def gather_deleted_ids(config):
-    '''
-    Connect to S3 and gather all of the transaction ids stored in CSV files
-    generated by the broker when transactions are removed from the DB.
-    '''
-    print('---------------------------------------------------------------')
-    if not config['provide_deleted']:
-        print('Skipping the S3 CSV fetch for deleted transactions')
-        return
-    print('Gathering all deleted transactions from S3')
-    start = perf_counter()
-    try:
-        s3 = boto3.resource('s3', region_name=config['aws_region'])
-        bucket = s3.Bucket(config['s3_bucket'])
-        bucket_objects = list(bucket.objects.all())
-    except Exception as e:
-        print('\n[ERROR]\n')
-        print('Verify settings.CSV_AWS_REGION and settings.DELETED_TRANSACTIONS_S3_BUCKET_NAME are correct')
-        print('  or is using env variables: CSV_AWS_REGION and DELETED_TRANSACTIONS_S3_BUCKET_NAME')
-        print('\n{}\n'.format(e))
-        raise SystemExit
-
-    if config['verbose']:
-        print('CSV data from {} to now.'.format(config['starting_date']))
-
-    to_datetime = datetime.combine(config['starting_date'], datetime.min.time(), tzinfo=pytz.UTC)
-    filtered_csv_list = [
-        x for x in bucket_objects
-        if (
-            x.key.endswith('.csv') and
-            not x.key.startswith('staging') and
-            x.last_modified >= to_datetime
-        )
-    ]
-
-    if config['verbose']:
-        print('Found {} csv files'.format(len(filtered_csv_list)))
-
-    deleted_ids = {}
-
-    for obj in filtered_csv_list:
-        # Use temporary files to facilitate date moving from csv files on S3 into pands
-        (file, file_path) = tempfile.mkstemp()
-        bucket.download_file(obj.key, file_path)
-
-        # Ingests the CSV into a dataframe. pandas thinks some ids are dates, so disable parsing
-        data = pd.DataFrame.from_csv(file_path, parse_dates=False)
-
-        if 'detached_award_proc_unique' in data:
-            new_ids = ['cont_tx_' + x for x in data['detached_award_proc_unique'].values]
-        elif 'afa_generated_unique' in data:
-            new_ids = ['asst_tx_' + x for x in data['afa_generated_unique'].values]
-        else:
-            print('  [Missing valid col] in {}'.format(obj.key))
-
-        # Next statements are ugly, but properly handle the temp files
-        os.close(file)
-        os.remove(file_path)
-
-        if config['verbose']:
-            print('{:<55} ... {}'.format(obj.key, len(new_ids)))
-
-        for uid in new_ids:
-            if uid in deleted_ids:
-                if deleted_ids[uid]['timestamp'] < obj.last_modified:
-                    deleted_ids[uid]['timestamp'] = obj.last_modified
-            else:
-                deleted_ids[uid] = {'timestamp': obj.last_modified}
-
-    if config['verbose']:
-        for k, v in deleted_ids.items():
-            print('id: {} last modified: {}'.format(k, str(v['timestamp'])))
-
-    print('Found {} IDs'.format(len(deleted_ids)))
-    print("Gathering deleted transactions took {} seconds".format(perf_counter() - start))
-    return deleted_ids
